+++ src/connectTool/index.js
@@ -20,16 +20,6 @@ export const makeLocaleProps = (state) => {
 };
 
 /**
- * Wraps a function with another function.
- * This is used for wrapping dispatch around a function.
- * @param {func} dispatch
- * @param {func} func
- * @return {func}
- */
-export const wrapFunc = (dispatch, func) => (...args) => dispatch(
-  func(...args));
-
-/**
  * This HOC initializes a store and locale for the tool.
  * It also specifies some required properties common to all tools.
  *
@@ -62,119 +52,14 @@ const connectTool = (namespace, options = {}) => {
     }
 
     const store = configureStore(reducer, middlewares);
-    // TRICKY: this will overwrite the default store context key
-    // thus removing direct access to tC core's store which also uses the default key.
-    // const Provider = createProvider();
 
-    // inject redux into the api and bind the lifecycle methods.
+    // wrap api in controller
     let controlledApi = undefined;
     if (api) {
       api.toString = () => namespace;
       controlledApi = new ApiController(api, store,
         hasLocale ? localeDir : undefined);
-      // , (state, dispatch, props) => {
-      //   // pre-process props before sending them to the tool
-      //   return {
-      //     ...props,
-      //     ...makeLocaleProps(state, hasLocale),
-      //     setToolLoading: wrapFunc(dispatch, setToolLoading),
-      //     setToolReady: wrapFunc(dispatch, setToolReady)
-      //   };
-      // });
     }
-
-    /**
-     * This container sets up the tool environment.
-     *
-     * @property {string} appLanguage - the app interface language code
-     */
-    // class Tool extends React.Component {
-    //   constructor(props) {
-    //     super(props);
-    //     this._isLoaded = this._isLoaded.bind(this);
-    //     this.handleChange = this.handleChange.bind(this);
-    //     this.state = {
-    //       broken: false,
-    //       error: null,
-    //       info: null
-    //     };
-    //   }
-    //
-    //   componentWillMount() {
-    //     // TODO: load this right away
-    //     const {appLanguage} = this.props;
-    //     if (hasLocale && !api) {
-    //       store.dispatch(loadLocalization(localeDir, appLanguage));
-    //     }
-    //     this.unsubscribe = store.subscribe(this.handleChange);
-    //   }
-    //
-    //   componentWillUnmount() {
-    //     this.unsubscribe();
-    //   }
-    //
-    //   handleChange() {
-    //     this.forceUpdate();
-    //   }
-    //
-    //   componentDidCatch(error, info) {
-    //     this.setState({
-    //       broken: true,
-    //       error,
-    //       info
-    //     });
-    //   }
-    //
-    //   componentWillReceiveProps(nextProps) {
-    //     // stay in sync with the application language
-    //     if (hasLocale && nextProps.appLanguage !== this.props.appLanguage) {
-    //       store.dispatch(setActiveLocale(nextProps.appLanguage));
-    //     }
-    //   }
-    //
-    //   /**
-    //    * Checks if the locale has finished loading
-    //    * @return {bool}
-    //    * @private
-    //    */
-    //   _isLoaded() {
-    //     const state = store.getState();
-    //     const toolIsLoaded = !isToolLoading(state);
-    //     const localeIsLoaded = !hasLocale || getLocaleLoaded(state);
-    //     return toolIsLoaded || localeIsLoaded;
-    //   }
-    //
-    //   render() {
-    //     const {broken, error, info} = this.state;
-    //     if(broken) {
-    //       return (
-    //         <BrokenScreen title="ERROR"
-    //                       error={error}
-    //                       info={info}/>
-    //       );
-    //     } else if (!this._isLoaded()) {
-    //       // TODO: we could display a loading screen while the tool loads
-    //       return null;
-    //     } else {
-    //       const localeProps = hasLocale ? makeLocaleProps(store.getState()) : {};
-    //       // TODO: rather than building the api prop hierarchy in tC we should pass flat props
-    //       // then here and in the api we can scope them to `tc`,
-    //       return (
-    //         <Provider store={store}>
-    //           <WrappedComponent
-    //             {...this.props}
-    //             toolApi={toolApi}
-    //             {...localeProps}
-    //           />
-    //         </Provider>
-    //       );
-    //     }
-    //   }
-    // }
-
-    /**
-     * This defines the interface between tools and tC core.
-     */
 
     return {
       name: namespace,
